portability and warning fixes
cobc:
* typeck.c (validate_alphabet): slightly rewritten to fix compiler warnings
* config.c, pparse.y (ppp_replace_list_add), reserved.c (is_invalid_word), tree.c, tree.h, typeck.c: minor adjustments to fix compiler warnings
* field.c (validate_multi_value, validate_elem_value), parser.y (setup_external_definition): fixed analyzer warnings by switching "for" (hint at "possibly NULL" at start) to "while"

libcob:
* intrinsic.c (split_around_t): explicit terminate snprintf
* termio.c (pretty_display_numeric): fix GCC warning with {{ 0 }}
* termio.c (display_alnum_dump): slightly rewritten to fix compiler warnings
* cconv.c (cob_load_collation), common.c, intrinisic.c (cob_intr_random), termio.c: minor adjustments to fix compiler warnings

+++ cobc/reserved.c
@@ -4144,7 +4144,7 @@ is_invalid_word (const char *word, const int size, const int space_allowed,
 
 	{
 		size_t i = 0;
-		while (i < size) {
+		while (i < (size_t)size) {
 			const char c = res_toupper(word[i++]);
 			if ((c >= 'A' && c <= 'Z')
 			 || (c >= '0' && c <= '9')
@@ -4152,7 +4152,7 @@ is_invalid_word (const char *word, const int size, const int space_allowed,
 				continue;
 			}
 			if (c == ' ' && space_allowed) {
-				if (i == 1 || word[i-2] != ' ') {
+				if (i == 1 || word[i - 2] != ' ') {
 					continue;
 				}
 			}
